COMP: Remove warnings

+++ Utilities/cmtar/compat/snprintf.c
@@ -527,7 +527,7 @@ static long round (long double value)
 {
   long intpart;
 
-  intpart = value;
+  intpart = (long)value;
   value = value - intpart;
   if (value >= 0.5)
     intpart++;
@@ -572,7 +572,7 @@ static void fmtfp (char *buffer, size_t *currlen, size_t maxlen,
   if (flags & DP_F_UP) caps = 1; /* Should characters be upper case? */
 #endif
 
-  intpart = ufvalue;
+  intpart = (long)ufvalue;
 
   /* 
    * Sorry, we only support 9 digits past the decimal because of our 
@@ -589,7 +589,7 @@ static void fmtfp (char *buffer, size_t *currlen, size_t maxlen,
   if (fracpart >= pow10 (max))
   {
     intpart++;
-    fracpart -= pow10 (max);
+    fracpart -= (long)(pow10 (max));
   }
 
 #ifdef DEBUG_SNPRINTF
